pyproject.toml
[tool.poetry]
name = "geodata-to-docx"
version = "0.1.0"
description = ""
authors = ["Danielle Gatland"]
license = "MIT License"
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = ">=3.9"
geopandas = "^0.14.0"
python-docx = "^1.1.0"
pillow = "^10.1.0"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
